refactor: migrate database layer to use sqlc and pgx/v5
* Replace sqlx with pgx/v5 for PostgreSQL interactions
* Update database queries to use sqlc for code generation
* Rename postgres package to sqlc
* Update habits repository to use new sqlc package
* Update response handling to use new response package
* Add sqlc configuration file
* Update go.mod and go.sum to reflect new dependencies
* Remove unused dependencies and code
* Add new files for sqlc generated code and response handling
* Update main function to use new database connection
* Update README to reflect new installation and configuration steps
* Update Makefile to check for sqlc installation
* Update .golangci.yml to skip sqlc generated code

+++ internal/habits/dto.go
@@ -8,11 +8,11 @@ type HabitWithLog struct {
 	ID          int32    `json:"id"`
 	Name        string   `json:"name"`
 	Description *string  `json:"description"`
-	LogValue    *float64 `json:"log_value"`
+	LogValue    *float32 `json:"log_value"`
 }
 
 type LogUpsertRequest struct {
 	HabitID int32     `json:"habit_id"`
 	Date    time.Time `json:"date"`
-	Value   float64   `json:"value"`
+	Value   float32   `json:"value"`
 }
